requirements.txt
# Core FastAPI dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
asyncpg==0.29.0
alembic==1.12.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# HTTP Clients & Async
httpx==0.25.2
aiohttp==3.9.1

# Validation & Serialization  
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# Blockchain & Crypto
base58==2.1.1

# Background Tasks (optional)
apscheduler==3.10.4

# Utilities
python-decouple==3.8
python-dotenv==1.0.0

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Production
gunicorn==21.2.0